rename instnace name
