+++ Cargo.toml
@@ -22,7 +22,7 @@ env_logger = "0.11.3"
 fs2 = "0.4.3"
 deeb_macros = "0.0.2"
 # deeb_core = "0.0.2"
-deeb_core = {path = "./deeb_core"}
+deeb_core = { path = "deeb_core" }
 
 [dev-dependencies]
 criterion = { version = "0.4", features = ["html_reports", "async_tokio"] }
