+++ src/lib.rs
@@ -186,7 +186,7 @@ mod deeb;
 pub use crate::deeb::Deeb;
 pub use deeb_core::{
     database::{
-        name::InstanceName, query::Query, transaction::Transaction, Database, ExecutedValue, Operation, DbResult
+        instance_name::InstanceName, query::Query, transaction::Transaction, Database, ExecutedValue, Operation, DbResult
     },
     entity::{Entity, EntityName},
 };
